pyproject.toml
[project]
name = "healthcare-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=38.2.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
